# openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "FIN-03 Agricultural Climate-Aware Credit Risk API",
    "version": "1.0.0",
    "description": "Proposed contract skeleton. Replace generic request/response placeholders with detailed schemas before implementation."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/auth/me": {
      "get": {
        "operationId": "getCurrentUser",
        "summary": "Get CurrentUser",
        "tags": [
          "MVP"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
    "/borrowers": {
      "get": {
        "operationId": "listBorrowers",
        "summary": "List Borrowers",
        "tags": [
          "MVP"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "post": {
        "operationId": "createBorrower",
        "summary": "Create Borrower",
        "tags": [
          "MVP"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/borrowers/{borrower_id}": {
      "get": {
        "operationId": "getBorrower",
        "summary": "Get Borrower",
        "tags": [
          "MVP"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "patch": {
        "operationId": "updateBorrower",
        "summary": "Update Borrower",
        "tags": [
          "MVP"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "borrower_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Unique borrower_id identifier"
        }
      ]
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}